requirements.txt
anyio==4.3.0
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.1.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
Unidecode==1.3.8
